+++ tests/testthat/test_table1.R
@@ -1,4 +1,3 @@
-library(furniture)
 library(testthat)
 
 test_that("table1 produces table1", {
@@ -96,14 +95,14 @@ test_that("table1 produces table1", {
   expect_s3_class(df %>% table1(c(1:3), splitby=~b, 
                          row_wise = TRUE,
                          test=TRUE, 
-                         output = "text2"), "data.frame")
+                         output = "text2"), "table1")
   expect_s3_class(df %>% table1(c(1:3), splitby=~b, 
                          test=TRUE, 
-                         output = "text2"), "data.frame")
+                         output = "text2"), "table1")
   expect_s3_class(df %>% table1(c(1:3), splitby=~b, 
                          row_wise = TRUE,
                          test=TRUE, 
-                         output = "latex"), "data.frame")
+                         output = "latex"), "knitr_kable")
   expect_s3_class(table1(df, c(1:3), splitby=~b, 
                          test=TRUE, 
                          output = "html"), "knitr_kable")